+++ CHANGES.md
@@ -50,6 +50,10 @@
 
   - <https://github.com/georust/gdal/pull/439>
 
+- Added histogram calculation
+
+  - <https://github.com/georust/gdal/pull/468>
+
 ## 0.16
 
 - **Breaking**: `Dataset::close` now consumes `self`